webserver/requirements.txt
python-dotenv
de-core-news-md @ https://github.com/explosion/spacy-models/releases/download/de_core_news_md-3.7.0/de_core_news_md-3.7.0-py3-none-any.whl
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0-py3-none-any.whl

faster-whisper==0.10.0
flask==3.0.0


Jinja2==3.1.2
joblib==1.3.2

ipykernel==6.26.0
ipython==8.12.3
jupyter-client==8.5.0
jupyter-core==5.5.0

matplotlib==3.7.4

mutagen==1.47.0

nltk==3.8.1
numpy==1.24.4

openai==1.5.0

pandas==2.0.3

playsound==1.3.0

protobuf==4.25.1


pydub==0.25.1

pyphen==0.14.0

requests==2.31.0

scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.2.2

sklearn==0.0

spacy==3.7.2


tabulate==0.9.0

torch==2.1.1
torchvision==0.16.1

tqdm==4.66.1

transformers==4.34.1

gunicorn


